Check users against a trust list
Check user cloaks or nicks against a trust list. If the user isn't
in the list, then error out or give the user a warning, based on
configuration. The trust list is a json object either generated
from a semantic mediawiki query, or via a raw json object that
comes from a configured URL.

Change-Id: I1a38629cbd8b6e4e244be9efa503e9d6e8048050