Enable logging to Twitter via Twitter API v1.1
Twitter disabled v1.0 of their API on 12 June 2013, at which point identi.ca's
cross-posting functionality ceased to work. identi.ca is migrating to a new
platform called 'pump.io' and their plans for re-enabling the Twitter bridge
are vague. On that date, #wikimedia-operation's morebots (which utilizes this
package) ceased to post updates to Twitter.

This patch allows adminlogbot to log directly to Twitter via the new API. It
does so using python-twitter v1+, which Faidon backported to Ubuntu Precise per
RT 5356. The patch adds detailed instructions for generating the requisite
access tokens to config.py.

I tested this change.

I am not a fan of the conditional import pattern or of the configuration file
being a Python module, but am following existing conventions in the interest of
seeing a fix through.

Someone in TechOps should follow up with another patch updating the metadata in
debian/ to declare the dependency on python-twitter and to bump the package
version.

Bug: 49476
RT: 5356
Change-Id: I18c836dea5d39aff4811780abdedd24a7865d3d7